feat(storybook): improve docs with MDX stories & also storybook config
