+++ .storybook/manager.js
@@ -1,0 +1,11 @@
+import { addParameters } from '@storybook/react';
+import { themes } from '@storybook/theming';
+
+addParameters({
+  options: {
+    name: 'UI components',
+    theme: themes.light,
+    showAddonPanel: false,
+    addonPanelInRight: true,
+  },
+});
