+++ .storybook/preview.js
@@ -1,0 +1,1 @@
+import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
