--- .storybook/webpack.config.js
@@ -1,17 +1,0 @@
-module.exports = async ({ config }) => {
-  config.module.rules.push({
-    test: /\.(ts|tsx)$/,
-    use: [
-      {
-        loader: require.resolve('awesome-typescript-loader'),
-      },
-      {
-        loader: require.resolve('react-docgen-typescript-loader'),
-      },
-    ],
-  });
-
-  config.resolve.extensions.push('.ts', '.tsx');
-
-  return config;
-};
