+++ stories/utils/items.ts
@@ -1,0 +1,22 @@
+export const menuItems = [
+  {
+    slug: '/#why',
+    label: 'Why',
+  },
+  {
+    slug: '/#what',
+    label: 'What',
+  },
+  {
+    slug: '/#how',
+    label: 'How',
+  },
+  {
+    slug: '/#success',
+    label: 'Success',
+  },
+  {
+    slug: '/#contact',
+    label: 'Contact',
+  },
+];
